Create spec for models.  Support ClojureScript.
In this commit, I've added some preliminary support for Clojure 1.9's
spec feature.  I feel like there's a lot more I could do with it, but
it's a start.  I'd definitely like to use generators in tests later.
It also might be nice to use generators to create random tours.

Also, Core and SA were converted to .cljc to support usage in
ClojureScript.  I'd like to use them in Teaspoon Demo, my Electron
project that I've been working on.  GA was not converted for two
reasons -- it doesn't work and it need more refactoring before it's
worth using.